+++ packages/iov-bns/src/bnsconnection.ts
@@ -6,8 +6,8 @@ import {
   AccountQuery,
   Address,
   AddressQuery,
+  AtomicSwap,
   AtomicSwapMerger,
-  BcpAtomicSwap,
   BcpAtomicSwapConnection,
   BcpSwapQuery,
   BcpTicker,
@@ -355,7 +355,7 @@ export class BnsConnection implements BcpAtomicSwapConnection {
   /**
    * All matching swaps that are open (from app state)
    */
-  public async getSwapsFromState(query: BcpSwapQuery): Promise<ReadonlyArray<BcpAtomicSwap>> {
+  public async getSwapsFromState(query: BcpSwapQuery): Promise<ReadonlyArray<AtomicSwap>> {
     const doQuery = (): Promise<QueryResponse> => {
       if (isQueryBySwapId(query)) {
         return this.query("/escrows", query.swapid);
@@ -380,7 +380,7 @@ export class BnsConnection implements BcpAtomicSwapConnection {
    *
    * To get claimed and returned, we need to look at the transactions.... TODO
    */
-  public async getSwaps(query: BcpSwapQuery): Promise<ReadonlyArray<BcpAtomicSwap>> {
+  public async getSwaps(query: BcpSwapQuery): Promise<ReadonlyArray<AtomicSwap>> {
     // we need to combine them all to see all transactions that affect the query
     const setTxs: ReadonlyArray<ConfirmedTransaction> = (await this.searchTx({
       tags: [bnsSwapQueryTags(query, true)],
@@ -414,7 +414,7 @@ export class BnsConnection implements BcpAtomicSwapConnection {
    *
    * This includes an open swap beind claimed/expired as well as a new matching swap being offered
    */
-  public watchSwaps(query: BcpSwapQuery): Stream<BcpAtomicSwap> {
+  public watchSwaps(query: BcpSwapQuery): Stream<AtomicSwap> {
     // we need to combine them all to see all transactions that affect the query
     const setTxs = this.liveTx({ tags: [bnsSwapQueryTags(query, true)] }).filter(isConfirmedTransaction);
     const delTxs = this.liveTx({ tags: [bnsSwapQueryTags(query, false)] }).filter(isConfirmedTransaction);